requirements.txt
tqdm==4.65.0
python-dotenv==1.0.0